% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkerboard.R
\name{checkerboard}
\alias{checkerboard}
\title{Checkerboard}
\usage{
checkerboard()
}
\description{
A demonstration of how the Multilevel Index of Dissimilarity measures spatial
clustering as well as unevenness
}
\details{
A criticism of the standard Index of Dissimilarity (ID) is that it only
measures one of the two principal dimensions of segregation - unevenness but
not spatial clustering. Because of this, very different spatial patterns of
segregation can generate the same ID score but the ID is unable to
distinguish between them.

In contrast, the multilevel index can detect the differences because
different patterns (scales) of segregation change the percentage of the
variance due to each level. The demonstation illustrates this using the
classic example of a checkerboard. The examples show how the percentage
of the total variance (labelled Pvariance) moves up the hierarchy with
the increase in spatial clustering at greater geographical cases. However,
the ID is always the same.

The 'stray' cell in examples 2-4 is to allow the model to be fitted. With it
the model correctly identifies that some of the variation remains at the
base level)
}
\seealso{
\code{\link{id}}
}
