% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo}
\alias{demo}
\title{Demonstration}
\usage{
demo()
}
\description{
A demonstration of how the Multilevel Index of Dissimilarity measures spatial clustering
}
\details{
A criticism of the standard Index of Dissimilarity (ID) is that it only measures one of the two
principle dimensions of segregation - unevenness but not spatial clustering. Because of this,
very different spatial patterns of segregation can generate the same ID score and so the ID is
unable to distinguish between them.

In contrast, the multilevel index can because different patterns (scales) of segregation
change the percentage of the variance due to each level. The demonstation provides some
examples.
}
