% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicities.R
\docType{data}
\name{ethnicities}
\alias{ethnicities}
\title{Population counts by ethnic group}
\format{A data frame with...}
\source{
Office for National Statistics; National Records of Scotland; Northern Ireland Statistics
and Research Agency (2016): 2011 Census aggregate data. UK Data Service (Edition: June 2016).
DOI: \url{http://dx.doi.org/10.5257/census/aggregate-2011-1}.
This information is licensed under the terms of the Open Government Licence
\url{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3}.
}
\usage{
ethnicities
}
\description{
Say more here
}
\keyword{datasets}
