% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLID.R
\docType{package}
\name{MLID}
\alias{MLID}
\alias{MLID-package}
\title{MLID: a package for calculating a multilevel index of dissimilarity}
\description{
Tools for fitting a multilevel index of dissimilarity to geographically hierarchical data.
The results measure the two principle dimensions of segregation, unevenness and clustering.
The amount of segregation, scale effects and the impact of particular places on the
overall index can be assessed.
}
