% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.index}
\alias{predict.index}
\title{Predicted index values}
\usage{
\method{predict}{index}(object, places, ...)
}
\arguments{
\item{places}{a character vector containing the names of the places in any of
the higher-level geographies for which the predictions will be made}

\item{...}{additional arguments (ignored)}

\item{index}{an object of class \code{index}: a multilevel index created
using function \code{\link{id}}}
}
\description{
Predicted values based on the contribution of chosen places to the overall
index
}
\details{
\code{predict.index} produces predicted values for the index of dissimilarity, obtained under
three scenarios. These are:
\itemize{
  \item if the (residual) effects for the chosen higher-level places are set
  to zero
  \item if the shares of the two population groups are equal everywhere
  except for the chosen places
  \item if all but the chosen places are omitted from the data and the index
  recalculated using only those that remain (it is the ID for the chosen
  places alone)
}
The predictions also include:
\itemize{
   \item the impact of the chosen places upon the overall ID, where a value
   over 100 indicates a group of neighbourhoods with a disproportionately
   high impact on the ID (in the same way that \code{\link{impacts}}
   calculates it)
   \item an R-squared value. This is the proportion of the total variation in
   (Y - X) that is due to the chosen places, where (Y - X) are the
   differences between the share of population group Y and the share of
   population group X that are observed in each neighbourhood.
}
}
\examples{
data("ethnicities")
index <- id(ethnicities, vars = c("Bangladeshi", "WhiteBrit"),
levels=c("LLSOA","MLSOA","LAD","RGN"))
ci <- confint(index)
catplot(ci)
# Note Tower Hamlets and Newham. Obtain the predictions for them:
predict(index, "Tower Hamlets")
predict(index, "Newham")
predict(index, c("Tower Hamlets","Newham"))
}
\seealso{
\code{\link{residuals.index}}
}
