% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{print.fxindex}
\alias{print.fxindex}
\title{Print values}
\usage{
\method{print}{fxindex}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{effect}}}

\item{...}{other arguments}
}
\description{
Prints predicted changes to the index of dissimilarity under various
scenarios
}
