% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts.R
\name{print.impacts}
\alias{print.impacts}
\title{Print values}
\usage{
\method{print}{impacts}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{impacts}}}

\item{...}{other arguments}
}
\description{
Prints the impact values
}
