% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printindex.R
\name{print.index}
\alias{print.index}
\title{Print values}
\usage{
\method{print}{index}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{id}}}

\item{...}{other arguments}
}
\description{
Prints output from the single or multi-level index of dissimilarity
}
\seealso{
\code{\link{id}} \code{\link{holdback}}
}
