% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{print.predictindex}
\alias{print.predictindex}
\title{Print values}
\usage{
\method{print}{predictindex}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{predict.index}}}

\item{...}{other arguments}
}
\description{
Prints predicted changes to the index of dissimilarity under various
scenarios generated by \code{\link{predict.index}}
}
