% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.index}
\alias{residuals.index}
\title{Extract Model Residuals}
\usage{
\method{residuals}{index}(index, ...)
}
\arguments{
\item{index}{an object of class \code{index}}
}
\value{
a numeric vector of matrix containing the residuals
}
\description{
For the Index of Dissimilarity (ID), the residuals are the differences between the
share of the Y population and the share of the X population per neighbourhood.
For the multilevel index, the residuals are estimated at each level of the model.
}
\details{
For the standard index, standardized (\code{rstandard}) and studentized (\code{rstudent})
residuals may also be calculated.
}
\examples{
data("ethnicities")
index <- id(ethnicities, vars = c("Bangladeshi", "WhiteBrit"))
resids <- rstandard(index)
summary(resids)
ethnicities[which.max(resids),]

index <- id(ethnicities, vars = c("Bangladeshi", "WhiteBrit"), levels=c("LLSOA","MLSOA","LAD","RGN"))
resids <- residuals(index)
head(resids)

## London is different from other regions:
sort(tapply(resids[,5], ethnicities$RGN, mean))

## At the local authority scale it is Tower Hamlets and Newham that have the highest share of the
## Bangladeshi population with respect to the White British:
tail(sort(tapply(resids[,4], ethnicities$LAD, mean)),5)
}
