% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{rstudent.index}
\alias{rstudent.index}
\title{The Studentised resdiduals for the single-level Index of Dissimilarity}
\usage{
\method{rstudent}{index}(model, ...)
}
\arguments{
\item{model}{an object of class \code{index} generated by the function
\code{\link{id}}}

\item{...}{other arguments}
}
\description{
Calculates the studentised residuals for the single-level index
}
\details{
The residuals are the differences between the share of the Y population
and the share of the X population per neighbourhood. A positive residual
occurs where the share of the Y population exceeds the share of the X
population, and a negative residual where the opposite. The studentised
residuals can help to identify 'extreme' differences.
}
