% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts.R
\name{tail.impacts}
\alias{tail.impacts}
\title{Return the lowest impact scores for each higher level area}
\usage{
\method{tail}{impacts}(x, n = 5, ...)
}
\arguments{
\item{x}{an object of class \code{impacts} generated by the function
\code{impacts}}

\item{n}{a single integer giving the number of rows to show. Defaults
to 5.}

\item{...}{other arguments}
}
\description{
Returns the last parts of the set of impact calculations, ordered
by the impact score
}
\seealso{
\code{\link{impacts}}
}
