% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.index}
\alias{residuals.index}
\title{Extract Model Residuals}
\usage{
\method{residuals}{index}(x, ...)
}
\arguments{
\item{x}{an object of class \code{index}}
}
\value{
a numeric vector of matrix containing the residuals
}
\description{
For the Index of Dissimilarity (ID), the residuals are the differences between the
share of the Y population and the share of the X population per neighbourhood.
For the multilevel index, the residuals are estimated at each level of the model.
}
\details{
For the standard index, standardized (\code{rstandard}) and studentized (\code{rstudent})
residuals may also be calculated.
}
\examples{
data("ethnicities")
index <- id(ethnicities, vars = c("Indian", "WhiteBrit"))
resids <- residuals(index)

index <- id(ethnicities, vars = c("Indian", "WhiteBrit"), levels=c("LLSOA","MLSOA","LAD","RGN"))
resids <- residuals(index)
head(resids)
## London is different from other regions:
sort(tapply(resids[,5], ethnicities$RGN, mean))
## But at a local authority scale it is Leicester that has the highest share of the
## Indian population with respect to the White British:
tail(sort(tapply(resids[,4], ethnicities$LAD, mean)),5)
}
\seealso{
\code{\link{id}}
}
